🗺️ Atlas: [architectural change]
🕸️ Tangle: `src/index/vector/temporal/mod.rs` had grown to over 3,700 lines due to thousands of lines of inline tests, creating an architectural "Bloat" anti-pattern that made the module hard to read and maintain.
📐 Blueprint: Extracted the inline `mod tests` block into a dedicated `src/index/vector/temporal/tests.rs` file and included it via `#[cfg(test)]\nmod tests;`.
🧱 Stability: Improved module cohesiveness by separating core implementation from test logic, reducing file size by ~2,300 lines while keeping test coverage intact.
🔬 Verification: Ran `cargo clippy` and `cargo test --lib index::vector::temporal` successfully. Added structural insight to `.jules/atlas.md`.
